+++ HW3/hw3_1/src/com/erofeev/array/runner/MainClass.java
@@ -1,0 +1,17 @@
+package com.erofeev.array.runner;
+
+import com.erofeev.array.arraytostring.ArraySort;
+
+public class MainClass {
+
+	public static void main(String[] args) {
+
+		ArraySort array = new ArraySort();
+
+		String[] testArray = { "Hello", "cruel", "World", "it", "ME", "BOB", "opop" };
+
+		System.out.println(array.findCapitalWords(testArray));
+
+	}
+
+}
