--- HW3/hw3_2/src/com/erofeev/hw3_1/ex4_runner/RandomGeneratorTest.java
@@ -1,15 +1,0 @@
-package com.erofeev.hw3_1.ex4_runner;
-
-import com.erofeev.hw3_2.ex4.NumberGenerator;
-
-public class RandomGeneratorTest {
-
-	public static void main(String[] args) {
-		
-		NumberGenerator  nubmergenerator = new NumberGenerator();
-	
-		nubmergenerator.sumDigit();
-
-	}
-
-}
