+++ HW3/hw3_2/src/com/erofeev/random/runner/NumberGeneratorTest.java
@@ -1,0 +1,15 @@
+package com.erofeev.random.runner;
+
+import com.erofeev.random.generator.NumberGenerator;
+
+public class NumberGeneratorTest {
+
+	public static void main(String[] args) {
+		
+		NumberGenerator  nubmergenerator = new NumberGenerator();
+	
+		nubmergenerator.sumDigit();
+
+	}
+
+}
