--- HW3/hw3_3/bin/.gitignore
@@ -1,4 +1,0 @@
-/com/
-/softwarecompany/
-/storehouse/
-/studio/
