--- HW3/hw3_3/src/com/erofeev/flowershop/flower/Chrysanthemum.java
@@ -1,15 +1,0 @@
-package com.erofeev.flowershop.flower;
-
-public class Chrysanthemum extends Flower {
-	private int freshness;
-
-	public Chrysanthemum(float price, int amount, String color, int freshness) {
-		super(price, amount, color);
-		this.freshness = freshness;
-	}
-
-	public float getPrice() {
-		return super.getPrice() * freshness;
-	}
-
-}
