--- HW3/hw3_3/src/com/erofeev/flowershop/flower/Lily.java
@@ -1,15 +1,0 @@
-package com.erofeev.flowershop.flower;
-
-public class Lily extends Flower {
-	private int freshness;
-
-	public Lily(float price, int amount, String color, int freshness) {
-		super(price, amount, color);
-		this.freshness = freshness;
-	}
-
-	public float getPrice() {
-		return super.getPrice() * freshness;
-	}
-
-}
