--- HW3/hw3_4/bin/.gitignore
@@ -1,1 +1,0 @@
-/com/
