--- HW3/hw3_4/src/com/erofeev/hw3_4/hospital/Pacient.java
@@ -1,12 +1,0 @@
-package com.erofeev.hw3_4.hospital;
-
-public class Pacient extends Person {
-	public Pacient(String firstName, String secondName, int age, boolean sex) {
-		super(firstName, secondName, age, sex);
-	}
-
-	public Pacient() {
-
-	}
-
-}
