+++ HW3/registrature/Hospital.png
