+++ HW3/registrature/Hospital.vpp
