+++ HW3/registrature/src/com/erofeev/hospital/entity/Patient.java
@@ -1,0 +1,12 @@
+package com.erofeev.hospital.entity;
+
+public class Patient extends Person {
+	public Patient(String firstName, String secondName, int age, boolean sex) {
+		super(firstName, secondName, age, sex);
+	}
+
+	public Patient() {
+
+	}
+
+}
